+++ src/cupcake/tofu/counting/chain_samples.py
@@ -8,6 +8,8 @@ from collections import OrderedDict, defaultdict
 from csv import DictReader, DictWriter
 from enum import Enum
 from multiprocessing import Process
+from typing import Union, List, Tuple
+from pathlib import Path
 
 import typer
 from Bio import SeqIO
@@ -15,8 +17,10 @@ from bx.intervals.cluster import ClusterTree
 
 from cupcake.sequence import GFF
 from cupcake.tofu.counting import combine_abundance_across_samples as sp
+from cupcake.logging import cupcake_logger as logger
 
-app = typer.Typer(name="chain_samples")
+
+app = typer.Typer(name="cupcake.tofu.counting.chain_samples")
 
 
 class fl_fields(str, Enum):
@@ -26,16 +30,13 @@ class fl_fields(str, Enum):
 
 def sample_sanity_check(
     group_filename, gff_filename, count_filename, fastq_filename=None
-):
+) -> None:
     """
     Double check that the formats are expected and all PBIDs are concordant across the files
     :return: raise Exception if sanity check failed
     """
-    print(
-        "Sanity checking. Retrieving PBIDs from {},{},{}...".format(
-            group_filename, gff_filename, count_filename
-        ),
-        file=sys.stderr,
+    logger.info(
+        f"Sanity checking. Retrieving PBIDs from {group_filename},{gff_filename},{count_filename}..."
     )
     ids1 = [line.strip().split()[0] for line in open(group_filename)]
     ids2 = [r.seqid for r in GFF.collapseGFFReader(gff_filename)]
@@ -51,18 +52,14 @@ def sample_sanity_check(
     ids3 = [r["pbid"] for r in DictReader(f, delimiter="\t")]
     if len(set(ids2).difference(ids1)) > 0 or len(set(ids2).difference(ids3)) > 0:
         raise Exception(
-            "Sanity check failed! Please make sure the PBIDs listed in {1} are also in {0} and {2}".format(
-                group_filename, gff_filename, count_filename
-            )
+            f"Sanity check failed! Please make sure the PBIDs listed in {gff_filename} are also in {group_filename} and {count_filename}"
         )
 
     if fastq_filename is not None:
         ids4 = [r.id.split("|")[0] for r in SeqIO.parse(open(fastq_filename), "fastq")]
         if len(set(ids2).difference(ids4)) > 0:
             raise Exception(
-                "Sanity check failed! Please make sure the PBIDs listed in {1} are also in {0}".format(
-                    fastq_filename, gff_filename
-                )
+                f"Sanity check failed! Please make sure the PBIDs listed in {gff_filename} are also in {fastq_filename}"
             )
 
 
@@ -90,34 +87,29 @@ def read_config(filename):
         for line in f:
             if line.startswith("tmpSAMPLE="):
                 if no_more_tmp:
-                    print(
-                        "Cannot have tmp_ samples after non-tmp_ samples! Abort!",
-                        file=sys.stderr,
+                    logger.error(
+                        "Cannot have tmp_ samples after non-tmp_ samples! Abort!"
                     )
                     sys.exit(-1)
                 name, path = line.strip()[len("tmpSAMPLE=") :].split(";")
                 if name.startswith("tmp_"):
-                    print(
-                        "Sample names are not allowed to start with tmp_! Please change {} to something else.".format(
-                            name
-                        ),
-                        file=sys.stderr,
+                    logger.error(
+                        f"Sample names are not allowed to start with tmp_! "
+                        f"Please change {name} to something else."
                     )
                     sys.exit(-1)
-                sample_dirs[name] = os.path.abspath(path)
-                sample_names.append("tmp_" + name)
+                sample_dirs[name] = Path(path).resolve()
+                sample_names.append(f"tmp_{name}")
             elif line.startswith("SAMPLE="):
                 no_more_tmp = True
                 name, path = line.strip()[len("SAMPLE=") :].split(";")
                 if name.startswith("tmp_"):
-                    print(
-                        "Sample names are not allowed to start with tmp_! Please change {} to something else.".format(
-                            name
-                        ),
-                        file=sys.stderr,
+                    logger.error(
+                        f"Sample names are not allowed to start with tmp_! "
+                        f"Please change {name} to something else."
                     )
                     sys.exit(-1)
-                sample_dirs[name] = os.path.abspath(path)
+                sample_dirs[name] = Path(path).resolve()
                 sample_names.append(name)
             elif line.startswith("GROUP_FILENAME="):
                 group_filename = line.strip()[len("GROUP_FILENAME=") :]
@@ -130,23 +122,19 @@ def read_config(filename):
 
     if group_filename is None:
         raise Exception(
-            "Expected GROUP_FILENAME= but not in config file {}! Abort.".format(
-                filename
-            )
+            f"Expected GROUP_FILENAME= but not in config file {filename}! Abort."
         )
     if count_filename is None:
         raise Exception(
-            "Expected COUNT_FILENAME= but not in config file {}! Abort.".format(
-                filename
-            )
+            f"Expected COUNT_FILENAME= but not in config file {filename}! Abort."
         )
     if gff_filename is None:
         raise Exception(
-            "Expected GFF_FILENAME= but not in config file {}! Abort.".format(filename)
+            f"Expected GFF_FILENAME= but not in config file {filename}! Abort."
         )
 
     if len(sample_names) == 0:
-        print("No samples given. Exit.", file=sys.stderr)
+        logger.error("No samples given. Exit.")
         sys.exit(-1)
 
     return (
@@ -159,20 +147,22 @@ def read_config(filename):
     )
 
 
-def read_count_info(count_filename, dirs, field_to_use):
+def read_count_info(
+    count_filename: Union[str, Path], dirs: List[Union[str, Path]], field_to_use: str
+) -> Tuple[str, str]:
     count_info = {}  # key: (sample, PB.1.1) --> count
     count_header = ""
     for name, d in dirs.items():
-        f = open(os.path.join(d, count_filename))
-        while True:
-            cur = f.tell()
-            line = f.readline().strip()
-            if not line.startswith("#"):
-                break
-            count_header += line
-        f.seek(cur)
-        for r in DictReader(f, delimiter="\t"):
-            count_info[name, r["pbid"]] = r[field_to_use]
+        with Path(d, count_filename).open() as f:
+            while True:
+                cur = f.tell()
+                line = f.readline().strip()
+                if not line.startswith("#"):
+                    break
+                count_header += line
+            f.seek(cur)
+            for r in DictReader(f, delimiter="\t"):
+                count_info[name, r["pbid"]] = r[field_to_use]
     return count_header, count_info
 
 
@@ -205,18 +195,16 @@ def chain_split_file(
     split_files = []
     i = 0
     counter = 0
-    f_gff = open(addon_gff + ".split" + str(i), "w")
-    f_group = open(addon_group + ".split" + str(i), "w")
+    f_gff = open(f"{addon_gff}.split{str(i)}", "w")
+    f_group = open(f"{addon_group}.split{str(i)}", "w")
     for v1 in tree.values():
         for strand in ("+", "-"):
             v2 = v1[strand]
-            for _start, _end, _indices in v2.getregions():
+            for *_, _indices in v2.getregions():
                 for cur in _indices:
                     GFF.write_collapseGFF_format(f_gff, recs[cur])
                     f_group.write(
-                        "{}\t{}\n".format(
-                            recs[cur].seqid, ",".join(addon_group_info[recs[cur].seqid])
-                        )
+                        f"{recs[cur].seqid}\t{','.join(addon_group_info[recs[cur].seqid])}\n"
                     )
                     counter += 1
             if counter >= (i + 1) * chunk_size:
@@ -231,8 +219,8 @@ def chain_split_file(
                     split_files.append((f_gff.name, f_group.name))
                 if i >= n_chunks or counter >= len(recs):
                     break
-                f_gff = open(addon_gff + ".split" + str(i), "w")
-                f_group = open(addon_group + ".split" + str(i), "w")
+                f_gff = open(f"{addon_gff}.split{str(i)}", "w")
+                f_group = open(f"{addon_group}.split{str(i)}", "w")
     if not f_gff.closed:
         n = f_gff.tell()
         f_gff.close()
@@ -254,7 +242,7 @@ def chain_split_file(
                 split_gff,
                 split_group,
                 ref_name,
-                addon_name + "." + str(i),
+                f"{addon_name}.{str(i)}",
                 fuzzy_junction,
                 allow_5merge,
                 max_3_diff,
@@ -262,26 +250,26 @@ def chain_split_file(
         )
         p.start()
         pools.append(p)
-        result_prefixes.append((ref_name, addon_name + "." + str(i)))
+        result_prefixes.append((ref_name, f"{addon_name}.{str(i)}"))
     for p in pools:
         p.join()
     return result_prefixes, split_files
 
 
 def chain_helper(
-    ref_gff,
-    ref_group,
-    addon_gff,
-    addon_group,
-    name1,
-    name2,
-    fuzzy_junction,
-    allow_5merge,
-    max_3_diff,
-):
+    ref_gff: Union[str, Path],
+    ref_group: Union[str, Path],
+    addon_gff: Union[str, Path],
+    addon_group: Union[str, Path],
+    name1: str,
+    name2: str,
+    fuzzy_junction: int,
+    allow_5merge: bool,
+    max_3_diff: int,
+) -> None:
     o = sp.MegaPBTree(
-        ref_gff,
-        ref_group,
+        gff_filename=ref_gff,
+        group_filename=ref_group,
         self_prefix=name1,
         internal_fuzzy_max_dist=fuzzy_junction,
         allow_5merge=allow_5merge,
@@ -289,10 +277,10 @@ def chain_helper(
         fastq_filename=None,
     )
     o.add_sample(
-        addon_gff,
-        addon_group,
+        gff_filename=addon_gff,
+        group_filename=addon_group,
         sample_prefix=name2,
-        output_prefix="tmp_" + name2,
+        output_prefix=f"tmp_{name2}",
         fastq_filename=None,
     )
 
@@ -322,9 +310,9 @@ def combine_split_chained_results(
     # sanity check files are all there
     split_files = []  # tuple of (gff, group, mega)
     for ref_name, o in output_prefixes:
-        gff_file = "tmp_" + o + ".gff"
-        mega_file = "tmp_" + o + ".mega_info.txt"
-        group_file = "tmp_" + o + ".group.txt"
+        gff_file = f"tmp_{o}.gff"
+        mega_file = f"tmp_{o}.mega_info.txt"
+        group_file = f"tmp_{o}.group.txt"
         if (
             not os.path.exists(gff_file)
             or not os.path.exists(mega_file)
@@ -458,15 +446,13 @@ def chain_samples(
         # we want to go to the last tmp_ sample and read it
         name = names[start_i - 1][4:]  # this is the last tmp_ sample, let's read it
         o = sp.MegaPBTree(
-            "tmp_" + name + ".gff",
-            "tmp_" + name + ".group.txt",
-            self_prefix="tmp_" + name,
+            f"tmp_{name}.gff",
+            f"tmp_{name}.group.txt",
+            self_prefix=f"tmp_{name}",
             internal_fuzzy_max_dist=fuzzy_junction,
             allow_5merge=allow_5merge,
             max_3_diff=max_3_diff,
-            fastq_filename="tmp_" + name + ".rep.fq"
-            if fastq_filename is not None
-            else None,
+            fastq_filename=f"tmp_{name}.rep.fq" if fastq_filename is not None else None,
         )
         # chain.append(name) # no need, already done above
     else:  # everything is new, start fresh
@@ -493,29 +479,27 @@ def chain_samples(
             os.path.join(d, gff_filename),
             os.path.join(d, group_filename),
             sample_prefix=name,
-            output_prefix="tmp_" + name,
+            output_prefix=f"tmp_{name}",
             fastq_filename=os.path.join(d, fastq_filename)
             if fastq_filename is not None
             else None,
         )
         o = sp.MegaPBTree(
-            "tmp_" + name + ".gff",
-            "tmp_" + name + ".group.txt",
-            self_prefix="tmp_" + name,
+            f"tmp_{name}.gff",
+            f"tmp_{name}.group.txt",
+            self_prefix=f"tmp_{name}",
             internal_fuzzy_max_dist=fuzzy_junction,
             allow_5merge=allow_5merge,
             max_3_diff=max_3_diff,
-            fastq_filename="tmp_" + name + ".rep.fq"
-            if fastq_filename is not None
-            else None,
+            fastq_filename=f"tmp_{name}.rep.fq" if fastq_filename is not None else None,
         )
         chain.append(name)
 
     # now recursively chain back by looking at mega_info.txt!!!
     d = {}  # ex: (tmp_1009, PB.1.1) --> mega info dict
     for c in chain[1:]:
-        for r in DictReader(open("tmp_" + c + ".mega_info.txt"), delimiter="\t"):
-            d["tmp_" + c, r["superPBID"]] = r
+        for r in DictReader(open(f"tmp_{c}.mega_info.txt"), delimiter="\t"):
+            d[f"tmp_{c}", r["superPBID"]] = r
 
     f1 = open("all_samples.chained_ids.txt", "w")
     writer1 = DictWriter(f1, fieldnames=["superPBID"] + chain, delimiter="\t")
@@ -524,7 +508,7 @@ def chain_samples(
     writer2 = DictWriter(f2, fieldnames=["superPBID"] + chain, delimiter="\t")
     writer2.writeheader()
 
-    reader = DictReader(open("tmp_" + chain[-1] + ".mega_info.txt"), delimiter="\t")
+    reader = DictReader(open(f"tmp_{chain[-1]}.mega_info.txt"), delimiter="\t")
     for r in reader:
         saw_NA = False
         r0 = r
@@ -536,11 +520,11 @@ def chain_samples(
         for c in chain[::-1][
             1:-1
         ]:  # the first sample does not have tmp_, because it's not a chain
-            if r["tmp_" + c] == "NA":
+            if r[f"tmp_{c}"] == "NA":
                 saw_NA = True
                 break
             else:
-                r2 = d["tmp_" + c, r["tmp_" + c]]
+                r2 = d[f"tmp_{c}", r[f"tmp_{c}"]]
                 answer[c] = r2[c]
                 if answer[c] != "NA":
                     answer2[c] = count_info[c, answer[c]]
@@ -560,9 +544,9 @@ def chain_samples(
     f1.close()
     f2.close()
 
-    shutil.copyfile("tmp_" + chain[-1] + ".gff", "all_samples.chained.gff")
+    shutil.copyfile(f"tmp_{chain[-1]}.gff", "all_samples.chained.gff")
     if fastq_filename is not None:
-        shutil.copyfile("tmp_" + chain[-1] + ".rep.fq", "all_samples.chained.rep.fq")
+        shutil.copyfile(f"tmp_{chain[-1]}.rep.fq", "all_samples.chained.rep.fq")
 
     print("Chained output written to:", file=sys.stdout)
     print("all_samples.chained.gff", file=sys.stdout)
@@ -627,10 +611,10 @@ def chain_samples_multithread(
                 else None
             )
         else:
-            ref_name = "tmp_" + ref_name
-            ref_gff = ref_name + ".gff"
-            ref_group = ref_name + ".group.txt"
-            ref_fq = ref_name + ".rep.fq" if fastq_filename is not None else None
+            ref_name = f"tmp_{ref_name}"
+            ref_gff = f"{ref_name}.gff"
+            ref_group = f"{ref_name}.group.txt"
+            ref_fq = f"{ref_name}.rep.fq" if fastq_filename is not None else None
         addon_d = dirs[addon_name]
         addon_gff = os.path.join(addon_d, gff_filename)
         addon_group = os.path.join(addon_d, group_filename)
@@ -654,7 +638,7 @@ def chain_samples_multithread(
 
         combine_split_chained_results(
             split_outs,
-            final_prefix="tmp_" + addon_name,
+            final_prefix=f"tmp_{addon_name}",
             ref_gff=ref_gff,
             ref_group=ref_group,
             ref_name=ref_name,
@@ -675,8 +659,8 @@ def chain_samples_multithread(
     # now recursively chain back by looking at mega_info.txt!!!
     d = {}  # ex: (tmp_sample1, PB.1.1) --> mega info dict
     for c in chain[1:]:
-        for r in DictReader(open("tmp_" + c + ".mega_info.txt"), delimiter="\t"):
-            d["tmp_" + c, r["superPBID"]] = r
+        for r in DictReader(open(f"tmp_{c}.mega_info.txt"), delimiter="\t"):
+            d[f"tmp_{c}", r["superPBID"]] = r
 
     f1 = open("all_samples.chained_ids.txt", "w")
     writer1 = DictWriter(f1, fieldnames=["superPBID"] + chain, delimiter="\t")
@@ -685,7 +669,7 @@ def chain_samples_multithread(
     writer2 = DictWriter(f2, fieldnames=["superPBID"] + chain, delimiter="\t")
     writer2.writeheader()
 
-    reader = DictReader(open("tmp_" + chain[-1] + ".mega_info.txt"), delimiter="\t")
+    reader = DictReader(open(f"tmp_{chain[-1]}.mega_info.txt"), delimiter="\t")
     for r in reader:
         saw_NA = False
         r0 = r
@@ -697,11 +681,11 @@ def chain_samples_multithread(
         for c in chain[::-1][
             1:-1
         ]:  # the first sample does not have tmp_, because it's not a chain
-            if r["tmp_" + c] == "NA":
+            if r[f"tmp_{c}"] == "NA":
                 saw_NA = True
                 break
             else:
-                r2 = d["tmp_" + c, r["tmp_" + c]]
+                r2 = d[f"tmp_{c}", r[f"tmp_{c}"]]
                 answer[c] = r2[c]
                 if answer[c] != "NA":
                     answer2[c] = count_info[c, answer[c]]
@@ -721,9 +705,9 @@ def chain_samples_multithread(
     f1.close()
     f2.close()
 
-    shutil.copyfile("tmp_" + chain[-1] + ".gff", "all_samples.chained.gff")
+    shutil.copyfile(f"tmp_{chain[-1]}.gff", "all_samples.chained.gff")
     if fastq_filename is not None:
-        shutil.copyfile("tmp_" + chain[-1] + ".rep.fq", "all_samples.chained.rep.fq")
+        shutil.copyfile(f"tmp_{chain[-1]}.rep.fq", "all_samples.chained.rep.fq")
 
     print("Chained output written to:", file=sys.stdout)
     print("all_samples.chained.gff", file=sys.stdout)