+++ CHANGELOG.md
@@ -4,7 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [19.1.1] - 2021-03-01
+
+### Changed
+- Most changes in cupcake.tofu.counting.chain_samples and 
+  cupcake.tofu.counting.chain_fusion_samples
+  - Replaced argparse with typer
+  - os.path with pathlib.Path
+  - format strings with f-strings
+  - Typing
 
 ## [19.1.0] - 2021-02-26